Difuzcam_requirements.txt
accelerate==0.27.0
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
datasets==2.17.0
diffusers==0.26.2
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.9.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.41
huggingface-hub==0.20.3
idna==3.4
ImageIO==2.37.2
importlib-metadata==7.0.1
Jinja2==3.1.2
lazy_loader==0.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.1
safetensors==0.4.2
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==1.40.3
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
tifffile==2025.5.10
tokenizers==0.15.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.2
transformers==4.37.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.4
urllib3==1.26.13
wandb==0.16.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
